pyproject.toml
[project]
name = "PTS"
version = "26.03.0"
description = "Open Targets Pipeline Transformation Stage"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
  "loguru==0.7.3",
  "opentargets-otter==26.3.4",
  "pandas>=2.3.2",
  "polars==1.38.1",
  "pyspark==3.5.7",
  "openai>=2.7.2",
  "defusedxml>=0.7.1",
  "openpyxl>=3.1.5",
  "setuptools>=80.9.0",
  "obonet>=0.3.1",
  "polars-hash>=0.5.6",
  "torch>=2.10.0",
  "ontoma==2.3.1",
  "transformers>=4.40.0,<5.0.0",
  "gcsfs>=2026.2.0",
  "pdfplumber>=0.11.9",
  "clinical-mining",
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cpu" }]
torchvision = [{ index = "pytorch-cpu" }]
clinical-mining = { git = "https://github.com/opentargets/clinical_mining" }


[dependency-groups]
dev = ["ruff==0.15.2", "ty==0.0.17", "deptry==0.24.0", "pre-commit>=4.3.0"]

[project.optional-dependencies]
test = [
  "coverage==7.13.4",
  "freezegun==1.5.5",
  "pytest==9.0.2",
  "pytest-asyncio==1.3.0",
  "pytest-mock==3.15.1",
]

[project.scripts]
pts = "pts.core:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.coverage.report]
exclude_lines = ["pragma: no cover", "if TYPE_CHECKING:"]

[tool.deptry]
pep621_dev_dependency_groups = ['test']
exclude = ["test"]

[tool.deptry.package_module_name_map]
opentargets-otter = 'otter'

[tool.deptry.per_rule_ignores]
DEP002 = ['openpyxl', 'coverage', 'setuptools', 'gcsfs']
DEP004 = ['pytest']

[tool.pytest.ini_options]
testpaths = ["test", "src"]
markers = ["slow: mark a test as slow (deselect with '-m \"not slow\"')"]
addopts = " --doctest-modules"

[tool.ruff]
target-version = 'py313'
line-length = 120
preview = true

[tool.ruff.format]
quote-style = 'single'

[tool.ruff.lint]
select = [
  'E',      # pycodestyle
  'F',      # pyflakes
  'N',      # pep8-naming
  'Q',      # flake8-quotes
  'RUF',    # ruff
  'UP',     # pyupgrade
  'D',      # pydocstyle
  'D212',   # flake8-docstrings
  'S',      # flake8-bandit
  'C4',     # flake8-comprehensions
  'INT',    # flake8-gettext
  'LOG',    # flake8-logging
  'G',      # flake8-logging-format
  'I',      # isort
  'ICN001', # flake8-import-conventions
  'B',      # flake8-bugbear
  'A001',   # flake8-builtins
  'COM',    # flake8-commas
  'T10',    # flake8-debugger
  'T20',    # flake8-todo
  'EXE',    # flake8-executable
  'ISC',    # flake8-implicit-str-concat
  'PIE',    # flake8-pie
  'PT',     # flake8-pytest-style
  'RSE',    # flake8-raise
  'RET504', # flake8-return
  'SIM',    # flake8-simplify
  'TID',    # flake8-tidy-imports
  'PGH',    # pygrep-hooks
  'PL',     # pylint
  'TRY',    # tryceratops
  'PERF',   # perflint
  'FURB',   # refurb
  'W',      # pycodestyle
]

ignore = [
  'D105',    # we don't want to document every single magic method
  'D107',    # we don't want to document every single init method
  'D213',    # we prefer D212
  'G004',    # we don't care about logging format
  'E226',    # allow omitting whitespace around arithmetic operators
  'E731',    # allow assigning lambdas (it's useful for single-line functions defined inside other functions)
  'RUF012',  # ultra-noisy and dicts in classvars are very common
  'RUF015',  # not always more readable
  'RUF067',  # we often put code inside init modules
  'S101',    # we use asserts outside tests, and do not run python with `-O` (also see B011)
  'S311',    # false positives, it does not care about the context
  'S324',    # all our md5/sha1 usages are for non-security purposes
  'S404',    # useless, triggers on *all* subprocess imports
  'S403',    # there's already a warning on using pickle, no need to have one for the import
  'S405',    # we don't use lxml in unsafe ways
  'S603',    # useless, triggers on *all* subprocess calls: https://github.com/astral-sh/ruff/issues/4045
  'S607',    # we trust the PATH to be sane
  'B011',    # we don't run python with `-O` (also see S101)
  'B904',    # possibly useful but too noisy
  'COM812',  # trailing commas on multiline lists are nice, but we have 2.5k violations
  'PIE807',  # `lambda: []` is much clearer for `load_default` in schemas
  'PT011',   # very noisy
  'PT015',   # nice for tests but not so nice elsewhere
  'PT018',   # ^ likewise
  'S301',    # we trust our pickled data
  'SIM102',  # sometimes nested ifs are more readable
  'SIM103',  # sometimes this is more readable (especially when checking multiple conditions)
  'SIM105',  # try-except-pass is faster and people are used to it
  'SIM108',  # noisy ternary
  'SIM114',  # sometimes separate ifs are more readable (especially if they just return a bool)
  'SIM117',  # nested context managers may be more readable
  'PLC0415', # local imports are there for a reason
  'PLC2701', # some private imports are needed
  'PLR09',   # too-many-<whatever> is just noisy
  'PLR0913', # very noisy
  'PLR1702', # complexity might be warranted
  'PLR2004', # extremely noisy and generally annoying
  'PLR6201', # sets are faster (by a factor of 10!) but it's noisy and we're in nanoseconds territory
  'PLR6301', # extremely noisy and generally annoying
  'PLW0108', # a lambda often makes it more clear what you actually want
  'PLW1510', # we often do not care about the status code of commands
  'PLW1514', # we expect UTF8 environments everywhere
  'PLW1641', # false positives with SA comparator classes
  'PLW2901', # noisy and reassigning to the loop var is usually intentional
  'TRY002',  # super noisy, and those exceptions are pretty exceptional anyway
  'TRY003',  # super noisy and also useless w/ werkzeugs http exceptions
  'TRY300',  # kind of strange in many cases
  'TRY301',  # sometimes doing that is actually useful
  'TRY400',  # not all exceptions need exception logging
  'PERF203', # noisy, false positives, and not applicable for 3.11+
  'FURB113', # less readable
  'FURB140', # less readable and actually slower in 3.12+
  'UP047',   # we need to support Python 3.11 for now
]

extend-safe-fixes = [
  'RUF005',  # we typically don't deal with objects overriding `__add__` or `__radd__`
  'C4',      # they seem pretty safe
  'UP008',   # ^ likewise
  'D200',    # ^ likewise
  'D400',    # ^ likewise
  'PT014',   # duplicate test case parametrizations are never intentional
  'RSE102',  # we do not use `raise func()` (with `func` returning the exception instance)
  'RET504',  # looks pretty safe
  'SIM110',  # ^ likewise
  'PERF102', # ^ likewise
]

[tool.ruff.lint.flake8-quotes]
inline-quotes = 'single'
multiline-quotes = 'double'
docstring-quotes = 'double'
avoid-escape = true

[tool.ruff.lint.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false
parametrize-names-type = 'tuple'
parametrize-values-type = 'list'
parametrize-values-row-type = 'tuple'

[tool.ruff.lint.per-file-ignores]
'test/*.py' = ['S108', 'S602', 'D']
'**/__init__.py' = ['F401']
'**/tasks/*.py' = ['D102']
'**/transformers/*.py' = ['D100', 'D103']
'**/pyspark/*.py' = ['RUF069', 'D100', 'D101', 'D102', 'D103']

[tool.ruff.lint.pydocstyle]
convention = "google"
